refactor(subscriptions): module divided - put filter, time_parser in seperate files - included mysql_test
